Refactor component imports and create a new Navbar component for improved structure
