+++ src/app/List_a_cause/See_Preview/page.tsx
@@ -3,7 +3,7 @@
 import React, { useEffect, useState } from "react";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
-import Navbar from "../component/navbar";
+import Navbar from "../_components/navbar";
 import { collection, addDoc } from "firebase/firestore";
 import { db } from "@/lib/firebase/config";
 import {
