+++ src/app/List_a_cause/page.tsx
@@ -1,49 +1,7 @@
-"use client";
-
-import React from "react";
-import Image from "next/image";
-import ListacauseForm from "./component/ListacauseForm";
-import Navbar from "./component/navbar";
+import ListacauseForm from "./_components/ListacauseForm";
 
 export default function Listacause() {
   return (
-    <div>
-      <Navbar />
-      <div className="w-full mx-auto md:flex md:flex-row">
-        <div className="hidden lg:block w-5/12 md:px-14 md:py-20 px-4 py-4 bg-[#f3f7fc]">
-          <div className="mb-80">
-            <h3 className="text-[#2b2829] text-xl font-semibold font-montserrat">
-              Welcome to the RefreeG cause listing!
-            </h3>
-            <p className="text-[#2b2829] text-base font-montserrat">
-              Let&rsquo;s get your cause started :)
-            </p>
-          </div>
-          <div>
-            <p className="mb-6 text-[#0a0a0b] text-base font-medium font-montserrat">
-              &quot;Every act of giving brings us closer to a brighter, more united
-              future. Your support isn&apos;t just a donation—it&apos;s a catalyst for
-              change&quot;
-            </p>
-            <div className="flex gap-3 items-center">
-              <Image
-                src="/List_a_cause/UserImg.svg"
-                alt=""
-                width={50}
-                height={50}
-              />
-              <span>
-                <h4>Unoagu Amadike Nomso</h4>
-                <p>Developer, Refreeg</p>
-              </span>
-            </div>
-          </div>
-        </div>
-
-        <div className="mx-auto lg:w-7/12 md:px-14 md:py-20 px-4 py-4">
-          <ListacauseForm />
-        </div>
-      </div>
-    </div>
+    <ListacauseForm />
   );
 }
